Lasers and Laser Calculations
8 hours of work later... Added a LaserModel that extends the standard
Model class that uses a line for calculations but renders a rectangle
for texture purposes. Added the ReflectionCalculation class to handle
the calculation of the possible intersections with all objects in the
scene and then sort accordingly until a final destination for a laser is
found, then move the laser's end point to that destination. Still to be
worked on: calculating reflection angle and then spawning a new laser
object to have its path calculated again. Also, textures could use a
work, we need a better system to allow multiple textures loaded and
applied to various models.