Lasers - Better but not Good
Angle of reflection is still broken, but Lasers have moved to vector
form, with an angle in radians and a length instead of beginning and end
coordinates. Also, the lag issues seem to be fixed, (knock on wood), but
that will be investigated later. Reflections are calculated correctly in
some cases, but not all of them; for example, see both lasers currently
in the scene.