Fixed Memory Issues and Rendering Issues
So, while in the middle of fixing the lasers, it became evident that
being able to see things would be nice. So a short workaround for the
non-rendering bug was fixed by running the generated loop infinitely,
but that brought back memory issues, which were fixed by deleting the
buffers before rewriting to them in the loop. This is a crude fix, but
Matt, you can move stuff around and sort out the tautological code.

This commit includes some changes to the Laser organization into a
wrapper, but is not functional in any way; that will come in a later
commit once finished.