WOOOO, Finally Finished Reflection
Finished refactoring the lasers into a new wrapper class that handles
the chain of reflections and rendering of all of them. Then made tweaks
to the reflection calculations to fix recursive reflections that were
added to the wrapper class. As it turns out, most issues were because
the initial point of the laser was valid as a final intersection point,
so the laser would never reflect away from the side of an object. Also
cleaned up a bit of code, though there are still some vestigal functions
that can be removed later.