GUI Button Implementation
Setup Button objects and all the code needed to setup non-ModelObject
renders on the screen and created callbacks for click registration on
the buttons. Next up is text objects and other, more complex, GUI
objects. Also fixed the scaling of the Camera so that the cartesian
coordinates in the game are from -10 to 10 instead of -12.5 to 12.5.