Movin' and Groovin'
Ok, just moving, no groovin' going on here. But there's now a new
MovableModel class that allows models to be dragged around with the
mouse, and it works beautifully. Also started organizing the UI code
with interfaces and a package. More UI code will come now that draggable
is done.