Added Triangles
Changed all of the code that was hard coded for 4 sides into variable
side numbers. Model now carries an integer of the number of sides it
has, and a triangle model creator was made. This should be easily
expandable up to n sides, but the math will get harder for declaring the
vertices based on xOffset and yOffset.