Updated Lasers and Clicking
Moved LaserStart over to an array to allow for multiple lasers in the
future. Fixed detection of mouse clicks that allowed the user to move
models around while the level was over. Removed unnecessary code in the
level classes for reflection of the lasers. Also updated the class
diagram because lol.