Code Cleanup
Added function to GameColor enum to return a floatColor instead of
needing to declare a new object inline. Removed theunused fullscreen
code. Moved GameColor to the render package. Attempted a second fix for
the reflectioncalculation concurrentModificationException crash, a
trycatch might be needed later if this doesn't solve it. Also removed a
lot of unused imports.