Bugs Be Gone
Fixed #2, Fixed #3. Added rotation methods to the laser wrapper to
rotate the laser along with a rotation of the starting block. For #2,
the movement was intended, as it centered the triangle on the cursor,
but the center point has been moved inside the triangle now to fix
issues with clicking.