Menus Abound
Lo! What's that on the horizon?! Is that a completely new menu system
developed from the ground up in only a few hours? Is that also a bunch
of new tools for UI creation including Text objects and fixed scaling?
OF COURSE IT IS!!!! Next step is to actually store the options to a text
file.