Ugliest Fix Ever
Fixed #9. The issue with edge detection arrises when the slope is
practically 0. It's fixed if the slope is adjusted to .01. This solves
every issue noticed so far, but has the side affect of making the line
thicker when reflecting off of a perpendicular slope (might be a good
thing to cue to the player that the laser is being reflected back). I've
since noticed an issue with lasers not reflecting a second time off of
movable blocks coming from immovable blocks, but I think we should just
ignore it.