Added Mod Loader
New Mod Loader class that will search for jars in the mods folder and
execute the constructor on the class with the same name as the jar file.
Jar classes will have access to everything in the game, but there isn't
a Mod API or useful functions for adding levels yet. See Test.jar and
Test.java for brief example.